Assignment Week 2
I created k means clustering plot based on remuneration and spending score. Five distinct plots were shown and it showed that groups that had a high or lower remuneration were more likely to spend or not spend on the extreme side. Those with a middle income spend a middle amount of the products. However a survey would be helpful in understanding what the customer motivation is for buying or not buying a game.